refactor(runner-pi): move AppstrateToolCtx into @appstrate/runner-pi
Make the type that defines `ctx.providerCall` (the credentialed-call
surface for custom tools, introduced in the previous commit) discoverable
from a public package.

`@appstrate/runner-pi` is already published on npm (publishConfig.access:
"public") and already exports related primitives like `readProviderRefs`
and `buildProviderCallExtensionFactory`. Putting `AppstrateToolCtx` next
to them lets tool authors `bun add -d @appstrate/runner-pi` and import
both `readProviderRefs` and the type from the same package, no new package
to publish.

Changes:
- New file `packages/runner-pi/src/tool-context.ts` — canonical home for
  `AppstrateToolCtx` and `AppstrateCtxProvider`.
- `packages/runner-pi/src/index.ts` — re-export the types.
- `runtime-pi/extension-wrapper.ts` — drop the local declarations, import
  `AppstrateCtxProvider` directly from `@appstrate/runner-pi`.
- `runtime-pi/entrypoint.ts` — import `AppstrateToolCtx` directly from
  `@appstrate/runner-pi` (alongside `readProviderRefs`), drop the import
  through the local wrapper module.

No behavior change. Tests still pass (11/11 in extension-wrapper.test.ts,
57/57 in runtime-pi/test).
