+++ runtime-pi/entrypoint.ts
@@ -56,8 +56,8 @@ import type { ProviderResolver } from "@appstrate/afps-runtime/resolvers";
 import type { ExecutionContext, RunEvent } from "@appstrate/afps-runtime/types";
 import { emptyRunResult } from "@appstrate/afps-runtime/runner";
 import { createMcpHttpClient, type AppstrateMcpClient } from "@appstrate/mcp-transport";
-import { wrapExtensionFactory, type AppstrateToolCtx } from "./extension-wrapper.ts";
-import { readProviderRefs } from "@appstrate/runner-pi";
+import { wrapExtensionFactory } from "./extension-wrapper.ts";
+import { readProviderRefs, type AppstrateToolCtx } from "@appstrate/runner-pi";
 import { attachStdoutBridge } from "@appstrate/afps-runtime/sinks";
 import { parseRuntimeEnv, RuntimeEnvError } from "./env.ts";
 import { buildMcpDirectFactories } from "./mcp/direct.ts";
